# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: dp-ring v2 Control API
  version: 0.1.0
  summary: Control-plane REST and SSE surface for operator-web.
servers:
  - url: http://127.0.0.1:7400
paths:
  /healthz:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Service health.
  /api/bootstrap:
    get:
      operationId: getBootstrap
      parameters:
        - $ref: "#/components/parameters/DpUserHeader"
        - $ref: "#/components/parameters/DpRoleHeader"
      responses:
        "200":
          description: Viewer, navigation, status catalog, and board counters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BootstrapPayload"
  /api/orgs:
    get:
      operationId: listOrgs
      responses:
        "200":
          description: Organizations visible to the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
  /api/orgs/{id}:
    get:
      operationId: readOrganization
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual organization identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/workspaces:
    get:
      operationId: listWorkspaces
      responses:
        "200":
          description: Workspaces visible to the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
  /api/workspaces/{id}:
    get:
      operationId: readWorkspace
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual workspace identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/repositories:
    get:
      operationId: listRepositories
      responses:
        "200":
          description: Repositories bound to workspaces.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Repository"
  /api/repositories/{id}:
    get:
      operationId: readRepository
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual repository identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/objectives:
    get:
      operationId: listObjectives
      responses:
        "200":
          description: Objectives in the active workspace.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Objective"
  /api/objectives/{id}:
    get:
      operationId: readObjective
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual objective identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Objective"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/blueprints:
    get:
      operationId: listBlueprints
      responses:
        "200":
          description: Execution blueprints for objectives.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Blueprint"
  /api/blueprints/{id}:
    get:
      operationId: readBlueprint
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual blueprint identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Blueprint"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/work-items:
    get:
      operationId: listWorkItems
      responses:
        "200":
          description: Queue and execution units.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkItem"
  /api/work-items/{id}:
    get:
      operationId: readWorkItem
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual work item identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkItem"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/executions:
    get:
      operationId: listExecutions
      responses:
        "200":
          description: Execution attempts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Execution"
  /api/executions/{id}:
    get:
      operationId: readExecution
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual execution identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Execution"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/reviews:
    get:
      operationId: listReviews
      responses:
        "200":
          description: Pending and completed reviews.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
  /api/reviews/{id}:
    get:
      operationId: readReview
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual review identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/publication-roots:
    get:
      operationId: listPublicationRoots
      responses:
        "200":
          description: Publication roots and their published members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicationRoot"
  /api/publication-roots/{id}:
    get:
      operationId: readPublicationRoot
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual publication root identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicationRoot"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/validation-reports:
    get:
      operationId: listValidationReports
      responses:
        "200":
          description: Validation reports generated for governance artifacts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationReport"
  /api/validation-reports/{id}:
    get:
      operationId: readValidationReport
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual validation report identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationReport"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/validation-results:
    get:
      operationId: listValidationResults
      responses:
        "200":
          description: Individual validation results emitted by a report run.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationResult"
  /api/validation-results/{id}:
    get:
      operationId: readValidationResult
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual validation result identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/findings:
    get:
      operationId: listFindings
      responses:
        "200":
          description: Findings and incidents.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Finding"
  /api/findings/{id}:
    get:
      operationId: readFinding
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual finding identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Finding"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/insights:
    get:
      operationId: listInsights
      responses:
        "200":
          description: Published learnings.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
  /api/insights/{id}:
    get:
      operationId: readInsight
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual insight identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/workers:
    get:
      operationId: listWorkers
      responses:
        "200":
          description: Registered worker daemons.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
  /api/workers/{id}:
    get:
      operationId: readWorker
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      responses:
        "200":
          description: Individual worker identified by id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /api/activity:
    get:
      operationId: listActivity
      responses:
        "200":
          description: Activity log snapshot.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
  /api/activity/stream:
    get:
      operationId: streamActivity
      responses:
        "200":
          description: Server-sent event stream of domain activity.
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Each emitted server-sent event frame is generated from an `ActivityEvent`.
                  - `id:` mirrors `ActivityEvent.id`
                  - `event:` mirrors `ActivityEvent.kind`
                  - `data:` contains the JSON-serialized `ActivityEvent` payload
                example: |-
                  id: evt_001
                  event: objective.submitted.v1
                  data: {"id":"evt_001","timestamp":"2026-04-17T00:00:00Z","kind":"objective.submitted.v1","title":"Objective submitted","summary":"Rebuild operator inbox entered triage.","resource_type":"objective","resource_id":"obj_inbox_rebuild"}
              x-dp-keepalive-seconds: 10
components:
  parameters:
    DpUserHeader:
      name: x-dp-user
      in: header
      required: false
      description: Optional viewer display-name override applied to the bootstrap payload.
      schema:
        type: string
    DpRoleHeader:
      name: x-dp-role
      in: header
      required: false
      description: Optional viewer role override for the bootstrap payload. Unknown values fall back to admin.
      schema:
        type: string
        enum: [admin, operator, reviewer, worker]
    ResourceIdPath:
      name: id
      in: path
      required: true
      description: Stable resource identifier from the seeded platform snapshot.
      schema:
        type: string
  responses:
    ResourceNotFound:
      description: No snapshot-backed resource matched the supplied id.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/NotFoundError"
  schemas:
    BootstrapPayload:
      type: object
      required: [viewer, navigation, status_catalog, boards]
      properties:
        viewer:
          $ref: "#/components/schemas/Viewer"
        navigation:
          type: array
          items:
            $ref: "#/components/schemas/NavSurface"
        status_catalog:
          type: array
          items:
            $ref: "#/components/schemas/StatusCatalog"
        boards:
          $ref: "#/components/schemas/Boards"
    Viewer:
      type: object
      required: [id, display_name, role, organization_id, workspace_id]
      properties:
        id: { type: string }
        display_name: { type: string }
        role: { type: string }
        organization_id: { type: string }
        workspace_id: { type: string }
    NavSurface:
      type: object
      required: [id, label, path, description]
      properties:
        id: { type: string }
        label: { type: string }
        path: { type: string }
        description: { type: string }
    StatusCatalog:
      type: object
      required: [entity, statuses]
      properties:
        entity: { type: string }
        statuses:
          type: array
          items: { type: string }
    Boards:
      type: object
      required: [objective_board, queue_board, run_timeline, review_inbox, worker_health]
      properties:
        objective_board: { $ref: "#/components/schemas/CounterMap" }
        queue_board: { $ref: "#/components/schemas/CounterMap" }
        run_timeline: { $ref: "#/components/schemas/CounterMap" }
        review_inbox: { $ref: "#/components/schemas/CounterMap" }
        worker_health: { $ref: "#/components/schemas/CounterMap" }
    CounterMap:
      type: object
      additionalProperties:
        type: integer
    NotFoundError:
      type: object
      required: [error, resource_id]
      properties:
        error:
          type: string
          enum: [resource_not_found]
        resource_id:
          type: string
    Organization:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
    Workspace:
      type: object
      required: [id, organization_id, name, slug, environment]
      properties:
        id: { type: string }
        organization_id: { type: string }
        name: { type: string }
        slug: { type: string }
        environment: { type: string }
    Repository:
      type: object
      required: [id, workspace_id, name, default_branch, clone_url, bound_worker_ids]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        name: { type: string }
        default_branch: { type: string }
        clone_url: { type: string }
        bound_worker_ids:
          type: array
          items: { type: string }
    Objective:
      type: object
      required: [id, workspace_id, title, summary, status, priority, repository_id, blueprint_id, active_work_item_ids, updated_at]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        title: { type: string }
        summary: { type: string }
        status: { type: string }
        priority: { type: string }
        repository_id: { type: string }
        blueprint_id: { type: string }
        active_work_item_ids:
          type: array
          items: { type: string }
        updated_at: { type: string, format: date-time }
    Blueprint:
      type: object
      required: [id, objective_id, status, summary, steps]
      properties:
        id: { type: string }
        objective_id: { type: string }
        status: { type: string }
        summary: { type: string }
        steps:
          type: array
          items:
            $ref: "#/components/schemas/BlueprintStep"
    BlueprintStep:
      type: object
      required: [id, title, summary]
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
    WorkItem:
      type: object
      required: [id, objective_id, blueprint_id, title, status, owner_lane, acceptance_contract, scope]
      properties:
        id: { type: string }
        objective_id: { type: string }
        blueprint_id: { type: string }
        title: { type: string }
        status: { type: string }
        owner_lane: { type: string }
        acceptance_contract:
          type: array
          items: { type: string }
        scope:
          $ref: "#/components/schemas/ScopeContract"
    ScopeContract:
      type: object
      required: [repository_id, target_path, file_allowlist, cleanup_paths]
      properties:
        repository_id: { type: string }
        target_path: { type: string }
        file_allowlist:
          type: array
          items: { type: string }
        build_command:
          type: [string, "null"]
        cleanup_paths:
          type: array
          items: { type: string }
    Execution:
      type: object
      required: [id, work_item_id, worker_id, status, launched_at, progress_percent, artifact_manifest]
      properties:
        id: { type: string }
        work_item_id: { type: string }
        worker_id: { type: string }
        status: { type: string }
        launched_at: { type: string, format: date-time }
        completed_at:
          type: [string, "null"]
          format: date-time
        progress_percent: { type: integer }
        artifact_manifest:
          type: array
          items:
            $ref: "#/components/schemas/ArtifactManifest"
    ArtifactManifest:
      type: object
      required: [kind, uri]
      properties:
        kind: { type: string }
        uri: { type: string }
    Review:
      type: object
      required: [id, work_item_id, execution_id, status, verdict, reviewer, notes]
      properties:
        id: { type: string }
        work_item_id: { type: string }
        execution_id: { type: string }
        status: { type: string }
        verdict: { type: string }
        reviewer: { type: string }
        notes: { type: string }
    GovernanceLocator:
      type: object
      properties:
        path: { type: string }
        kind: { type: string }
        id: { type: string }
        branch_id: { type: string }
        checkpoint_id: { type: string }
    GovernanceDescriptor:
      type: object
      required: [name, mediaType, digest, size, locator]
      properties:
        name: { type: string }
        mediaType: { type: string }
        digest: { type: string }
        size: { type: integer }
        locator:
          $ref: "#/components/schemas/GovernanceLocator"
    PublicationMemberDescriptor:
      allOf:
        - $ref: "#/components/schemas/GovernanceDescriptor"
        - type: object
          required: [artifact_type]
          properties:
            profile:
              type: [string, "null"]
            artifact_type: { type: string }
    PublicationRoot:
      type: object
      required: [_type, id, conforms_to, status, about, member_descriptors, membership_digest]
      properties:
        _type: { type: string }
        id: { type: string }
        conforms_to: { type: string }
        status: { type: string }
        about:
          $ref: "#/components/schemas/GovernanceDescriptor"
        member_descriptors:
          type: array
          items:
            $ref: "#/components/schemas/PublicationMemberDescriptor"
        membership_digest: { type: string }
    GovernanceSubjectRef:
      type: object
      required: [type, id]
      properties:
        type: { type: string }
        id: { type: string }
    ValidationReportSummary:
      type: object
      required: [info, warning, violation]
      properties:
        info: { type: integer }
        warning: { type: integer }
        violation: { type: integer }
    ValidationReportData:
      type: object
      required: [subject_ref, profile_id, report_level, conforms, outcome, result_ids, summary]
      properties:
        subject_ref:
          $ref: "#/components/schemas/GovernanceSubjectRef"
        profile_id: { type: string }
        report_level: { type: string }
        conforms: { type: boolean }
        outcome: { type: string }
        result_ids:
          type: array
          items: { type: string }
        summary:
          $ref: "#/components/schemas/ValidationReportSummary"
    ValidationReport:
      type: object
      required: [id, type, version, created_at, updated_at, created_by, status, data]
      properties:
        id: { type: string }
        type: { type: string, enum: [validation-report] }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string }
        session_id:
          type: [string, "null"]
        status: { type: string }
        data:
          $ref: "#/components/schemas/ValidationReportData"
    ValidationResultData:
      type: object
      required: [report_id, subject_ref, subject_location, rule_id, rule_location, severity, detail_result_ids]
      properties:
        report_id: { type: string }
        subject_ref:
          $ref: "#/components/schemas/GovernanceSubjectRef"
        subject_location: { type: string }
        rule_id: { type: string }
        rule_location: { type: string }
        severity: { type: string }
        message:
          type: [string, "null"]
        detail_result_ids:
          type: array
          items: { type: string }
    ValidationResult:
      type: object
      required: [id, type, version, created_at, updated_at, created_by, status, data]
      properties:
        id: { type: string }
        type: { type: string, enum: [validation-result] }
        version: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        created_by: { type: string }
        session_id:
          type: [string, "null"]
        status: { type: string }
        data:
          $ref: "#/components/schemas/ValidationResultData"
    Finding:
      type: object
      required: [id, workspace_id, title, status, severity, source_execution_id]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        title: { type: string }
        status: { type: string }
        severity: { type: string }
        source_execution_id: { type: string }
    Insight:
      type: object
      required: [id, workspace_id, title, status, category, excerpt]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        title: { type: string }
        status: { type: string }
        category: { type: string }
        excerpt: { type: string }
    Worker:
      type: object
      required: [id, workspace_id, display_name, status, capacity, active_executions, repository_ids, capabilities]
      properties:
        id: { type: string }
        workspace_id: { type: string }
        display_name: { type: string }
        status: { type: string }
        capacity: { type: integer }
        active_executions: { type: integer }
        repository_ids:
          type: array
          items: { type: string }
        capabilities:
          type: array
          items: { type: string }
    ActivityEvent:
      type: object
      required: [id, timestamp, kind, title, summary, resource_type, resource_id]
      properties:
        id: { type: string }
        timestamp: { type: string, format: date-time }
        kind: { type: string }
        title: { type: string }
        summary: { type: string }
        resource_type: { type: string }
        resource_id: { type: string }